+++ src/Utilities/DerivativeMatrix.cpp
@@ -1,36 +1,36 @@
 #include <functional>
 #include <cstring>
-#include "../../include/Utilities/LagrangePolynomials.h"
-#include "../../include/Utilities/PolyEval.h"
-#include "../../include/Utilities/PolyDeriv.h"
-#include "../../include/Utilities/LobattoIntegration.h"
-#include "../../include/Utilities/MassMatrix.h"
+#include "../../includes/Utilities/LagrangePolynomials.h"
+#include "../../includes/Utilities/PolyEval.h"
+#include "../../includes/Utilities/PolyDeriv.h"
+#include "../../includes/Utilities/LobattoIntegration.h"
+#include "../../includes/Utilities/MassMatrix.h"
 
 using namespace std;
 
-void derivativeMatrix(float *DerivativeMatrix,unsigned N)
+void derivativeMatrix(double *DerivativeMatrix,unsigned N)
 {
-    float Poly[N+1][N+1];
-    float **poly, **deriv;
-    poly   =   new float*[N+1];
-    deriv  =   new float*[N+1];
+    double Poly[N+1][N+1];
+    double **poly, **deriv;
+    poly   =   new double*[N+1];
+    deriv  =   new double*[N+1];
     lagrangePolynomials(*Poly,N);
     unsigned i,j;
 
     for(i=0;i<=N;i++)
     {
-        poly[i] =   new float[N+1];
-        deriv[i]=   new float[N];
-        memcpy(poly[i],Poly[i],(N+1)*sizeof(float));
+        poly[i] =   new double[N+1];
+        deriv[i]=   new double[N];
+        memcpy(poly[i],Poly[i],(N+1)*sizeof(double));
         polyDeriv(poly[i],deriv[i],N+1);
     }
 
-    function<float(float)> eval;
+    function<double(double)> eval;
     for(i=0;i<=N;i++)
     {
         for(j=0;j<=N;j++)
         {
-            eval = [&poly,&deriv,&i,&j,&N](float x){return (((polyEval(poly[i],N,x))*(polyEval(deriv[j],N-1,x))));};
+            eval = [&poly,&deriv,&i,&j,&N](double x){return (((polyEval(poly[i],N,x))*(polyEval(deriv[j],N-1,x))));};
             DerivativeMatrix[i*(N+1)+j] = lobattoIntegration(-1.0,1.0,N+1,eval);
         }
     }
@@ -46,10 +46,10 @@ void derivativeMatrix(float *DerivativeMatrix,unsigned N)
     return ;
 }
 
-void twoDDerivativeMatrixX(float *DerivativeMatrix, unsigned N)
+void twoDDerivativeMatrixX(double *DerivativeMatrix, unsigned N)
 {
-    float m[N+1][N+1];
-    float d[N+1][N+1];
+    double m[N+1][N+1];
+    double d[N+1][N+1];
     derivativeMatrix(*d,N);
     massMatrix(*m,N);
 
@@ -63,10 +63,10 @@ void twoDDerivativeMatrixX(float *DerivativeMatrix, unsigned N)
     return ;
 }
 
-void twoDDerivativeMatrixY(float *DerivativeMatrix, unsigned N)
+void twoDDerivativeMatrixY(double *DerivativeMatrix, unsigned N)
 {
-    float m[N+1][N+1];
-    float d[N+1][N+1];
+    double m[N+1][N+1];
+    double d[N+1][N+1];
     derivativeMatrix(*d,N);
     massMatrix(*m,N);
