+++ src/Utilities/Ones.cpp
@@ -1,11 +1,13 @@
-void ones(float *x, unsigned n)
+#include "../../includes/Utilities/Ones.h"
+
+void ones(double *x, unsigned n)
 {
     unsigned i;
     for(i=0; i<n; i++)
         x[i]    =   1.0;
 }
 
-void ones(float *A, unsigned m, unsigned n)
+void ones(double *A, unsigned m, unsigned n)
 {
     unsigned i;
     for(i=0;i<m*n;i++)
