+++ src/Utilities/Plot.cpp
@@ -2,10 +2,11 @@
 #include <iostream>
 #include <string>
 #include <cstdlib>
-#include "../../include/Utilities/Display.h"
+#include "../../includes/Utilities/Display.h"
+#include "../../includes/Utilities/Plot.h"
 using namespace std;
 
-void plot(float *X, float *Y, float N, string GraphTitle = "", string LegendTitle = "", string fileName = "")
+void plot(double *X, double *Y, double N, string GraphTitle = "", string LegendTitle = "", string fileName = "")
 {
     unsigned i;
     FILE* pfile;
@@ -37,7 +38,7 @@ void plot(float *X, float *Y, float N, string GraphTitle = "", string LegendTitl
 /**
   * A similar 1-D plot but with the y-axis ranges.
   */
-void plot(float *X, float *Y, float N,float Ymin, float Ymax, string GraphTitle = "", string LegendTitle = "", string fileName = "")
+void plot(double *X, double *Y, double N,double Ymin, double Ymax, string GraphTitle = "", string LegendTitle = "", string fileName = "")
 {
     unsigned i;
     FILE* pfile;
@@ -73,7 +74,7 @@ void plot(float *X, float *Y, float N,float Ymin, float Ymax, string GraphTitle 
  * @param Y    The Y-Domain.
  * @param Z    The Z-Domain.
  */
-void plot(float *X, float *Y, float *Z, unsigned Nx, unsigned Ny,float X1, float X2,float Y1, float Y2,float Z1, float Z2, string Name)
+void plot(double *X, double *Y, double *Z, unsigned Nx, unsigned Ny,double X1, double X2,double Y1, double Y2,double Z1, double Z2, string Name)
 {
     int i;
     FILE* pfile;
