+++ src/Utilities/PolyDeriv.cpp
@@ -1,4 +1,6 @@
-void polyDeriv(float *GivenPoly, float *DerivedPoly, unsigned deg)
+#include "../../includes/Utilities/PolyDeriv.h"
+
+void polyDeriv(double *GivenPoly, double *DerivedPoly, unsigned deg)
 {
 	for(int i = 1;i<=deg;i++)
         DerivedPoly[i-1] = i*GivenPoly[i];
